Remove saved state for user pokemon
Removal of saved state for user pokemon will remove some of the
problems and features that are tied to overall gameplay of Pokemon
games rather than solely battling, and therefore will make focusing
on just the battling aspect easier.

An example of such a problem is the evolution of Pokemon: this isn't
a fundamental part of the mechanics of battling yet it almost always
occurs during a battle. However, tackling the problems associated with
incorporating Pokemon evolution, while interesting, are not immediately
relevant for battling.